แก้ไขไฟล์

+++ form_show.php
@@ -1,3 +1,17 @@
+<?php 
+$db_name = "test"; //ชื่อฐานข้อมูล
+$db_host = "localhost";
+$db_user = "root"; //ชื่อuser
+$db_pass = "1234@#+"; //ชื่อรหัสผ่าน
+
+try {
+   $db_con = new PDO( "mysql:host={$db_host};dbname={$db_name}", $db_user, $db_pass );
+   $db_con->setAttribute( PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION );
+   $db_con->exec( "set names utf8" );
+} catch ( PDOException $e ) {
+   echo $e->getMessage();
+}
+?>
 <!DOCTYPE html>
 <html lang="en">
 
@@ -12,7 +26,7 @@
 <body>
    <div class="container">
       <div class="row justify-content-center">
-         <div class="col-md-7 py-5">
+         <div class="col-md-8 py-5">
             <h2 class="pb-0">ฟอร์มแสดงข้อมูลสมาชิก</h2>
             <nav aria-label="breadcrumb" class="mb-5">
                <ol class="breadcrumb">
@@ -27,41 +41,28 @@
                      <th>ลำดับที่</th>
                      <th>ชื่อ-สกุล</th>
                      <th>อีเมลล์</th>
-                     <th>สร้างเมื่อ</th>
+                     <th>ที่อยู่</th>
                      <th>จัดการ</th>
                   </tr>
                </thead>
                <tbody>
+                  <?php 
+                  $n = 1; 
+                  $stmt = $db_con->prepare("SELECT * FROM users ORDER BY id DESC");
+                  $stmt->execute();
+                  while ($rows = $stmt->fetch(PDO::FETCH_ASSOC)) {?>
                   <tr>
-                     <td scope="row">1</td>
-                     <td>Exp</td>
-                     <td>Exp</td>
-                     <td>Exp</td>
+                     <td scope="row"><?php echo $n;?></td>
+                     <td><?php echo $rows["name"];?> <?php echo $rows["surname"];?></td>
+                     <td><?php echo $rows["email"];?></td>
+                     <td><?php echo $rows["address"];?></td>
                      <td width="140" align="center">
                         <a href="#" class="btn btn-info" role="button">แก้ไข</a>
-                        <a href="./form_actions.php?do=delete" class="btn btn-danger" role="button">ลบ</a>
+                        <a href="./form_actions.php?do=delete&id=<?php echo $rows["id"];?>" onclick="return confirm('คุณต้องการลบข้อมูลแถวนี้ใช่หรือไม่ !!');" class="btn btn-danger" role="button">ลบ</a>
                      </td>
                   </tr>
-                  <tr>
-                     <td scope="row">2</td>
-                     <td>Exp</td>
-                     <td>Exp</td>
-                     <td>Exp</td>
-                     <td width="140" align="center">
-                        <a href="#" class="btn btn-info" role="button">แก้ไข</a>
-                        <a href="#" class="btn btn-danger" role="button">ลบ</a>
-                     </td>
-                  </tr>
-                  <tr>
-                     <td scope="row">3</td>
-                     <td>Exp</td>
-                     <td>Exp</td>
-                     <td>Exp</td>
-                     <td width="140" align="center">
-                        <a href="#" class="btn btn-info" role="button">แก้ไข</a>
-                        <a href="#" class="btn btn-danger" role="button">ลบ</a>
-                     </td>
-                  </tr>
+                  <?php $n++; }
+                  ?>
                </tbody>
             </table>
          </div>
